minos: Change `FileIdentity.volume_serial` from `u64` to `u32`
This is sufficient for storing both the `DWORD` used to identify a file's
storage device under windows, as well as the dev_t used under linux.

+++ infra/minos.hpp
@@ -151,7 +151,7 @@ namespace minos
 	
 	struct FileIdentity
 	{
-		u64 volume_serial;
+		u32 volume_serial;
 
 		u64 index;
 	};
